pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shayde"
version = "0.1.0"
description = "Docker Playwright E2E testing and screenshot capture CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Shayde Team" }
]
keywords = ["playwright", "e2e", "testing", "screenshot", "visual-regression", "docker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "typer>=0.12.0",
    "playwright>=1.48.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "aiohttp>=3.9.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
shayde = "shayde.cli.main:app"

[project.urls]
Homepage = "https://github.com/tatun55/shayde"
Repository = "https://github.com/tatun55/shayde"

[tool.hatch.build.targets.wheel]
packages = ["src/shayde"]

[tool.hatch.build.targets.sdist]
include = ["src/shayde", "tests"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
